Implement large_dnode pool feature
This feature adds support for variable length dnodes. The size may be a
multiple of 512 bytes up to the size of a dnode block (currently 16384
bytes).  A dn_nextra field was added to the current on-disk dnode_phys_t
structure to describe the size of the physical dnode on disk.  The 8
bits for this field were taken from the zero filled dn_pad2 field. The
field represents how many "extra" dnode_phys_t slots a dnode consumes in
its dnode block. This convention results in a value of 0 for 512 byte
dnodes which preserves on-disk format compatibility with older pools.

Similarly, the in-core dnode structure has a new dn_count field to
represent to total number of dnode_phys_t slots consumed on disk. Thus
dn->dn_count is 1 greater than the corresponding dnp->dn_nextra. This
difference in convention was adopted because, unlike on-disk structures,
backward compatibility is not a concern for in-core objects.  We were
free, therefore, to choose a more natural representation of size in a
dnode_t, rather than the awkward "nextra" used by dnode_phys_t for
compatibility reasons.

To make use of the new variable length dnode feature, the "dnodesize"
dataset property was introduced. When a new object is allocated, this
property is used to set the size of that object's dnode. By default the
property is set to 512 bytes, but that can be changed dynamically. Thus
dnodes of varying sizes can coexist within the same dataset and even
within the same dnode block. For example, to start using 1K dnodes for
all new objects, run the following command:

  # zfs set dnodesize=1k tank/fish

The size of dnodes created in a transaction is determined at transaction
creation time and stored as a new field in dmu_tx. This makes the value
easily available throughout the creation call path and ensures it does
not change for the life of the transaction (i.e. even if the dnodesize
dataset property is changed). dmu_tx_create() will take the value from
the current default stored in the objset. A new interface
dmu_tx_create_dn_count() is added to create a transaction with a
specified dnode count. This is for use during ZIL replay and
send/receive.

As a result of the now variable length size of a dnode on disk, various
parts of the code had to be updated:

* dnode_hold_impl's prototype was modified to take a "count"
  parameter. When the DNODE_MUST_BE_FREE flag is used, this
  parameter is used to ensure the "hole" at the specified object
  offset is large enough to hold the dnode being created.

  For example, consider a dnode meta object that has object 10 and 12
  allocated, but not 11. If a hold is taken on object 11 with flag
  DNODE_MUST_BE_FREE and sectors set to 2, we must fail. There is
  only a single 512b sector available at object offset 11 and we're
  requesting to hold a new dnode with size of 2 sectors. This 2
  sector large dnode cannot fit at object offset 11, so we fail.

  In addition, this new "count" parameter is used to ensure a
  dnode object does not span multiple dnode blocks.
  In both of these cases, if a failure occurs, ENOSPC is returned.
  Keep in mind, these failure cases are only possible when using the
  DNODE_MUST_BE_FREE flag.

  If the DNODE_MUST_BE_ALLOCATED flag is
  set, "count" must be 0. dnode_hold_impl() will check that the
  requested dnode is not already consumed as an extra dnode by
  an earlier large dnode.

* dnode_next_offset_level() and other functions that iterate
  through blocks filled with dnodes may no longer use a simple array
  indexing scheme. These now use the current dnode's dn_count
  field to iterate to the next dnode in the block. This is to ensure
  we properly skip the current dnode's bonus area and don't
  interpret it as a valid dnode.

Finally, the zdb command was updated to display a dnode's dn_szsec
field when the object is dumped under the "szsec" column.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Signed-off-by: Ned Bass <bass6@llnl.gov>