WIP prototype tiny ZAP (experimental)
Add a new type of ZAP that allows up to 8 64-bit integers to be stored
in the value but still fits in a single block. This is a proposed
optimization for Lustre since it want to store a 128-bit FID in the
directory entry which currently requires a fat ZAP.

Add a new dataset property "zap_default_type" which can be set to one of
"micro", "tiny", or "fat".  The default value is micro which preserves
normal ZFS behavior. Setting the property to tiny or fat will force ZFS
to use those ZAP types by default.

TODO:

- automatically upgrade microzaps to tinyzaps when an entry being
  added meets the criteria.

- update *_uint64 ZAP functions to support tinyzaps

- support tiny zaps with > 128k block sizes

Signed-off-by: Ned Bass <bass6@llnl.gov>