Backfill metadnode more intelligently
Only attempt to backfill lower metadnode object numbers if at least 4096
objects have been freed since the last rescan, and at most once per
transaction group. This avoids a pathology in dmu_object_alloc() that
caused O(N^2) behavior for create-heavy workloads and substantially
improves object creation rates. As summarized by @mahrens in #4636:

"Normally, the object allocator simply checks to see if the next
object is available. The slow calls happened when dmu_object_alloc()
checks to see if it can backfill lower object numbers. This happens
every time we move on to a new L1 indirect block (i.e. every 32 *
128 = 4096 objects).  When re-checking lower object numbers, we use
the on-disk fill count (blkptr_t:blk_fill) to quickly skip over
indirect blocks that don’t have enough free dnodes (defined as an L2
with at least 393,216 of 524,288 dnodes free). Therefore, we may
find that a block of dnodes has a low (or zero) fill count, and yet
we can’t allocate any of its dnodes, because they've been allocated
in memory but not yet written to disk. In this case we have to hold
each of the dnodes and then notice that it has been allocated in
memory.

The end result is that allocating N objects in the same TXG can
require CPU usage proportional to N^2."

Signed-off-by: Ned Bass <bass6@llnl.gov>